backend-edited-recorded-run-model

+++ backend/cadence/models/RecordedRunModel.py
@@ -7,9 +7,9 @@ class RecordedRun(models.Model):
     recorded_on = models.DateField()
     start_time = models.TimeField()
     end_time = models.TimeField()
-    real_pace = models.IntegerField(validators=[MinValueValidator(0)])
-    real_distance = models.IntegerField(validators=[MinValueValidator(0)])
-    real_duration = models.IntegerField(validators=[MinValueValidator(0)])
+    real_pace = models.FloatField(validators=[MinValueValidator(0)])
+    real_distance = models.FloatField(validators=[MinValueValidator(0)])
+    real_duration = models.FloatField(validators=[MinValueValidator(0)])
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     run = models.ForeignKey(Run, on_delete=models.CASCADE, null=True)
